+++ celery-codegen/src/lib.rs
@@ -1,7 +1,5 @@
 #![recursion_limit = "256"]
 
-extern crate proc_macro;
-
 use proc_macro::TokenStream;
 
 mod error;